Revert to using ASP.NET bundler.
Use ASP.NET bundler on vendor libraries, angular dependencies and angular app files.
Use gulp bundler to moving assets and running bower install.
