--- Presentation.Web/Scripts/notify/notify.js
@@ -1,1 +1,0 @@
-angular.module('notify', []);
