+++ Presentation.Web/Views/Shared/_Layout.cshtml
@@ -9,10 +9,11 @@
 
     <link href="//kendo.cdn.telerik.com/2016.1.112/styles/kendo.common-bootstrap.min.css" rel="stylesheet" type="text/css" />
     <link href="//kendo.cdn.telerik.com/2016.1.112/styles/kendo.bootstrap.min.css" rel="stylesheet" type="text/css" />
-    <link href="/Content/css/app.min.css" rel="stylesheet"/>
+    @Styles.Render("~/Content/css/css")
 
-    <script src="/Scripts/library-bundle.min.js"></script>
-    <script src="/Scripts/angular-bundle.min.js"></script>
+    @Scripts.Render("~/Scripts/vendor")
+    @Scripts.Render("~/Scripts/angular")
+
     <script src="//kendo.cdn.telerik.com/2016.1.112/js/jszip.min.js"></script>
     <script src="//kendo.cdn.telerik.com/2016.1.112/js/kendo.web.min.js"></script>
     <!-- Kendo localization -->
