+++ Presentation.Web/app/shared/notify/notify.module.ts
@@ -1,0 +1,1 @@
+angular.module("notify", []);
