+++ package.json
@@ -30,9 +30,7 @@
     "gulp-minify-css": "^1.2.3",
     "gulp-open": "^1.0.0",
     "gulp-protractor": "^1.0.0",
-    "gulp-sourcemaps": "^1.6.0",
     "gulp-tslint": "^3.5.0",
-    "gulp-uglify": "^1.5.1",
     "gulp-util": "^3.0.7",
     "jasmine-core": "^2.3.4",
     "jasmine-expect": "^1.22.3",
